Cast to double

+++ src/main/java/com/visma/of/rp/routeevaluator/constraintsAndObjectives/objectives/TimeWindowObjectiveFunction.java
@@ -4,6 +4,7 @@ public class TimeWindowObjectiveFunction extends CustomCriteriaObjectiveFunction
 
     public TimeWindowObjectiveFunction() {
         super(objectiveInfo -> !objectiveInfo.isDestination(),
-                objectiveInfo -> Math.max(0.0, objectiveInfo.getVisitEnd() - objectiveInfo.getTask().getEndTime()));
+                (objectiveInfo -> (double) Math.max(0, objectiveInfo.getVisitEnd() -
+                        objectiveInfo.getTask().getEndTime())));
     }
 }
